Return the Information class to its intended file
