+++ csfieldguide/static/interactives/packet-attack/css/packet-attack.scss
@@ -4,7 +4,7 @@
     padding: 20px;
 }
 
-@media screen and (min-width: 820px) and (min-height: 650px) {
+@media screen and (min-width: 820.1px) and (min-height: 650.1px) {
     .warning {
         display: none;
     }
