+++ csfieldguide/templates/interactives/packet-attack.html
@@ -4,7 +4,7 @@
 {% load static %}
 
 {% block html %}
-  <div class="mb-3 container-fluid">
+  <div class="my-3 container-fluid">
     <div class="row interactive game-canvas">
       <div class="col-12" id="interactive-packet-attack"></div>
     </div>
